Sequential outlier rejection for the AI, EBS, and GOA

+++ R/sor_save_results.R
@@ -7,23 +7,25 @@
 #' @param cruise_idnum One or more cruise ID numbers as a numeric vector (e.g. 757). Must be the same length as final_dir.
 #' @param survey Survey name prefix to use in output filename (e.g. NBS_2022)
 #' @param channel An RODBC channel. Will prompt user to get connected if NULL.
+#' @param delete_existing Logical indicating whether to overwrite data in RACE_DATA.EDIT_HAUL_IMPORT_SOR_UPDATES
 #' @import RODBC
 #' @export
 
-sor_save_results <- function(final_dir, create_user = "", survey, cruise_idnum, channel = NULL) {
-  
-  stopifnot("sor_save_results: cruise_idnum and final_dir must be the same length." = length(cruise_idnum) == length(final_dir))
-  
-  channel <- get_connected(schema = "AFSC")
+sor_save_results <- function(final_dir, 
+                             create_user = "", 
+                             survey, 
+                             cruise_idnum, 
+                             channel = NULL, 
+                             delete_existing = TRUE) {
   
   fpath <- character(length = 0L)
   
   for(jj in final_dir) {
-    fpath <- c(fpath, list.files(jj, 
-                                 pattern = "final.rds",
-                                 full.names = TRUE))
+    fpath <- c(fpath, 
+               list.files(jj, 
+                          pattern = "final.rds",
+                          full.names = TRUE))
   }
-
   
   final_values <- data.frame()
   
@@ -48,7 +50,7 @@ sor_save_results <- function(final_dir, create_user = "", survey, cruise_idnum, 
   final_values$CREATE_USER <- toupper(create_user)
 
   # Write output to .csv
-  csv_path <- here::here("output", paste0("race_data_edit_hauls_table_", survey, ".csv"))
+  csv_path <- here::here("output", paste0("race_data_edit_hauls_table_", survey[1], ".csv"))
   
   message(paste0("sor_save_results: Writing results to ", csv_path))
   
@@ -56,36 +58,51 @@ sor_save_results <- function(final_dir, create_user = "", survey, cruise_idnum, 
             file = csv_path, 
             row.names = FALSE)
   
+  channel <- get_connected(channel = channel, 
+                           schema = "AFSC")
   
   # Clear existing data from the table
-  delete_existing <- readline(paste0("Any existing data must be deleted from RACE_DATA.EDIT_HAUL_IMPORT_SOR_UPDATES before adding new results. Should data for cruise(s) ", paste(cruise_idnum, collapse = ", "), " be deleted (y or n)?"))
+  if(is.null(delete_existing)) {
+    
+    delete_existing <- readline(
+      paste0("Any existing data must be deleted from RACE_DATA.EDIT_HAUL_IMPORT_SOR_UPDATES before adding new results. Should data for cruise(s) ", 
+             paste(cruise_idnum, collapse = ", "), " be deleted (y or n)?"))
+    
+    delete_existing <- switch(tolower(delete_existing),
+                              "y" = TRUE,
+                              "n" = FALSE)
+    
+  }
   
-  delete_existing <- tolower(delete_existing)
-  
-  stopifnot("sor_save_results: Execution halted. User opted not to delete existing data from " = delete_existing == "y")
-
-  message("sor_save_results: Removing existing cruise data from RACE_DATA.EDIT_HAUL_IMPORT_SOR_UPDATES")
-  
-  RODBC::sqlQuery(channel = channel, 
-                  query = paste0(" DELETE FROM RACE_DATA.EDIT_HAUL_IMPORT_SOR_UPDATES WHERE CRUISE_ID IN (", paste(cruise_idnum, collapse = ", "), ");"))
-  
-  # Append data to table  
-  message("sor_save_results: Appending new data to RACE_DATA.EDIT_HAUL_IMPORT_SOR_UPDATES")
-
-  RODBC::sqlSave(channel = channel, 
-                 dat = final_values, 
-                 tablename = "RACE_DATA.EDIT_HAUL_IMPORT_SOR_UPDATES",
-                 append = TRUE, # need append = TRUE because of access permissions
-                 rownames = FALSE, 
-                 colnames = FALSE, 
-                 verbose = FALSE,
-                 safer = FALSE, 
-                 addPK = FALSE, 
-                 # typeInfo, 
-                 # varTypes,
-                 fast = TRUE, 
-                 test = FALSE, 
-                 nastring = NULL)
+  if(delete_existing) {
+    
+    message("sor_save_results: Removing existing cruise data from RACE_DATA.EDIT_HAUL_IMPORT_SOR_UPDATES")
+    
+    RODBC::sqlQuery(channel = channel, 
+                    query = paste0(" DELETE FROM RACE_DATA.EDIT_HAUL_IMPORT_SOR_UPDATES WHERE CRUISE_ID IN (", paste(cruise_idnum, collapse = ", "), ");"))
+    
+    # Append data to table  
+    message("sor_save_results: Appending new data to RACE_DATA.EDIT_HAUL_IMPORT_SOR_UPDATES")
+    
+    RODBC::sqlSave(channel = channel, 
+                   dat = final_values, 
+                   tablename = "RACE_DATA.EDIT_HAUL_IMPORT_SOR_UPDATES",
+                   append = TRUE, # need append = TRUE because of access permissions
+                   rownames = FALSE, 
+                   colnames = FALSE, 
+                   verbose = FALSE,
+                   safer = FALSE, 
+                   addPK = FALSE, 
+                   # typeInfo, 
+                   # varTypes,
+                   fast = TRUE, 
+                   test = FALSE, 
+                   nastring = NULL)
+  } else {
+    
+    message("sor_save_results: Execution halted. User opted not to delete existing data.")
+    
+  }
   
   return(final_values)
